+++ src/Synthesizer/Generator_Sine.cpp
@@ -1,5 +1,5 @@
 #include "Generator_Sine.h"
 
-inline double synthesizer::Generator_Sine::soundFunction(noteBuffer& noteBuffer){
+inline float synthesizer::Generator_Sine::soundFunction(noteBuffer& noteBuffer){
     return sin(noteBuffer.phaze*noteBuffer.multiplier);
 }
