+++ src/Synthesizer/Generator_Square.cpp
@@ -1,5 +1,5 @@
 #include "Generator_Square.h"
 
-inline double synthesizer::Generator_Square::soundFunction(noteBuffer& noteBuffer){
+inline float synthesizer::Generator_Square::soundFunction(noteBuffer& noteBuffer){
     return (int((noteBuffer.phaze)/noteBuffer.multiplier) & 0x1)*2 - 1;
 }
