+++ src/Synthesizer/settings.h
@@ -18,13 +18,13 @@ namespace synthesizer{
         dynamics attack;  //how long signal rises to maximum volume
         dynamics sustain; //how long signal stays at maximum volume
         dynamics fade;    //how long signal fades after sustain part even if note is still being played
-        double fadeTo;    //value fade strives to achieve
-        double rawFadeTo; //raw value set by user
+        float fadeTo;    //value fade strives to achieve
+        float rawFadeTo; //raw value set by user
         dynamics release; //how long signal is still being faded down after it is not being played anymore
 
         uint dynamicsDuration; //ovarall duration of attack, sustain and release
 
-        double stereoMix;
+        float stereoMix;
     };
 }
 #endif
